#138 separate composited widgets, fix flatAlert
